Refactor project structure and enhance functionality
- Updated package.json to change project name and description, and replaced `cp` with `copyfiles` for asset copying.
- Added `rimraf` for cleaning the `dist` directory.
- Updated TypeScript types for improved type safety in `SaveResponse`.
- Enhanced input validation in `tanaPayloadBuilder.ts` to ensure required parameters are provided.
- Improved error handling in `getStorageConfig` to reject promises on validation failure.
- Updated Vite target to support Chrome 135.
- Adjusted manifest.json to reflect project name and description changes.

These changes aim to improve code quality, maintainability, and user experience.
