+++ static/manifest.json
@@ -1,8 +1,8 @@
 {
   "manifest_version": 3,
-  "name": "Save to Tana (TypeScript)",
+  "name": "Save to Tana",
   "version": "1.1.0",
-  "description": "Save web page content to Tana with structured fields and supertags - TypeScript Edition",
+  "description": "Save web page content to Tana with structured fields and supertags.",
   "permissions": [
     "activeTab",
     "storage",
